Threaded and Notifier based playback
The original solution relied upon autonomousPeriodic being called every
cycle, and delaying, as necessary to get to the offset from start time
that will match the original driver/operator actions.  This solution
suffered from excessive jitter in playback timing.  The outward symptom
was that, when played back, the robot would end in positions that were
far (a few feet, in the worst cases) from the recorded run, or from a
set of playback runs.  This implementation still exists, at least in
part as playbackLong.

A second solution was tried which involved spinning up a thread which
would manage the timing using delay calls.  It appears that this timing
is still tied to JVM timing, and not to FPGA timing.  The jitter, in
most cases, is better, but there are still cases in which the playback
is far from the original recording.

The best solution, thus far, creates a Notifier object to manage the
timing.  The autonomousPeriodic call, the first time through, creates a
list iterator for the recorded driver inputs, and a Notifier. It
schedules an invocation of an internal class, AutoOperation.
AutonOperation implements runnable, which is a requirement for Notifier.
When run, AutoOperation invokes the method on the object assigned at
ActionRecorder creation, and then schedules the next event.

At this point, most runs happen with little jitter, and the jitter that
is present is a few milliseconds, and those events are sparse.  I
suspect that the problems come up when consecutive events are more than
10 ms late.